+++ src/core/server/systems/athenaEvent.ts
@@ -1,12 +1,19 @@
 import * as alt from 'alt-server';
-import { ATHENA_EVENTS_PLAYER, ATHENA_EVENTS_VEHICLE } from '../enums/athenaEvents';
+import {
+    ATHENA_EVENTS_PLAYER,
+    ATHENA_EVENTS_PLAYER_CLIENT,
+    ATHENA_EVENTS_VEHICLE
+} from '../../shared/enums/athenaEvents';
 
 const events: { [key: string]: Array<any> } = {};
 
 type playerCallback = (result: alt.Player, ...args: any) => void;
 type vehicleCallback = (result: alt.Vehicle, ...args: any) => void;
 
-function on(eventName: ATHENA_EVENTS_PLAYER | ATHENA_EVENTS_VEHICLE, callback: playerCallback | vehicleCallback) {
+function on(
+    eventName: ATHENA_EVENTS_PLAYER | ATHENA_EVENTS_VEHICLE | ATHENA_EVENTS_PLAYER_CLIENT,
+    callback: playerCallback | vehicleCallback
+) {
     if (!events[eventName]) {
         events[eventName] = [];
     }
@@ -23,11 +30,11 @@ export class EventController {
     /**
      * Subscribe to an Athena Player event.
      * @static
-     * @param {ATHENA_EVENTS_PLAYER} eventName
+     * @param {ATHENA_EVENTS_PLAYER | ATHENA_EVENTS_PLAYER_CLIENT} eventName
      * @param {playerCallback} callback
      * @memberof EventController
      */
-    static onPlayer(eventName: ATHENA_EVENTS_PLAYER, callback: playerCallback) {
+    static onPlayer(eventName: ATHENA_EVENTS_PLAYER | ATHENA_EVENTS_PLAYER_CLIENT, callback: playerCallback) {
         on(eventName, callback);
     }
 
@@ -61,6 +68,12 @@ Object.values(ATHENA_EVENTS_PLAYER).forEach((eventName) => {
 
 Object.values(ATHENA_EVENTS_VEHICLE).forEach((eventName) => {
     alt.on(eventName, (...args: any[]) => {
+        processCallbacks(eventName.toString(), args);
+    });
+});
+
+Object.values(ATHENA_EVENTS_PLAYER_CLIENT).forEach((eventName) => {
+    alt.onClient(eventName, (...args: any[]) => {
         processCallbacks(eventName.toString(), args);
     });
 });