+++ src/core/shared/enums/playerSynced.ts
@@ -2,5 +2,6 @@ export enum PLAYER_SYNCED_META {
     WANTED_LEVEL = 'player-wanted',
     POSITION = 'player-position',
     PING = 'player-ping',
-    NAME = 'player-name'
+    NAME = 'player-name',
+    WAYPOINT = 'player-waypoint'
 }